FIX: name in start and stop are misaligned
In the README, the container is started with `--name deepl-mock`. However, the container is stopped with name 'deepl/deepl-mock' leading to an cli error message of "No such container: deepl/deepl-mock".

In order to correct this misalignment, the docker stop command should be changed to stop the container with name 'deepl-mock'.
Changing the name to 'deepl/deepl-mock' is not possible, due to the character '/' not allowed in docker container names.